calc ang gen scripts

+++ phi4/graphs/calculate.py
@@ -4,6 +4,97 @@ import os
 import roperation
 import sympy
 import utils
+import subprocess
+import fnmatch
+import re as regex
+import time
+
+
+def execute(name, model, points=10000, threads=2, calc_delta=0., neps=0):
+    dirname = '%s/%s/'%(model.workdir,name)
+    MAXPOINTS=10**9
+    if points >= MAXPOINTS:
+        iterations = 2*int(points/MAXPOINTS)
+        points = MAXPOINTS
+    else:
+        iterations = 2
+    os.chdir(dirname)
+    filelist=os.listdir(".")
+    result=[]
+    error=[]
+    for n in range(neps+1):
+        result.append(0.)
+        error.append(0.)
+        for file in filelist:
+            if fnmatch.fnmatch(file,"*E%s_*.run"%n):
+                arg_lst=["./%s"%file, "%s"%points, "%s"%threads]
+                arg_lst.append(str(iterations))
+                if calc_delta<>0:
+                    arg_lst.append("%s"%calc_delta)
+                process = subprocess.Popen(arg_lst, shell=False, 
+                                            stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+                exit_code = process.wait()
+                (std_out,std_err) = process.communicate()
+                if exit_code<>0:
+                    raise Exception,  "Failed while executing %s"%file
+                res,err=parse_output(std_out)
+                if res==None or err==None:
+                    raise ValueError, "Failed while parsing %s output"%file
+                result[n]+=res
+                error[n]+=err
+    time_str=time.strftime("-%Y-%m-%d-%H:%M:%S")
+    f=open("result_%s"%time_str,'w')
+    f.write(str((result,error)))
+    f.close()
+    f=open("points_%s"%time_str,'w')
+    f.write(str(points))
+    f.close()
+    return (result, error)
+
+
+def parse_output(stdout):
+    res=None
+    err=None
+    for line in stdout.splitlines():
+        reg = regex.match("^result = (.+)$", line)
+        if reg:
+            if 'nan' not in reg.groups()[0]:
+                res = float(reg.groups()[0])
+        reg = regex.match("^std_dev = (.+)$", line)
+        if reg:
+            if 'nan' not in reg.groups()[0]:
+                err = abs(float(reg.groups()[0]))
+    return (res, err)
+
+def compile(name,model):
+    dirname = '%s/%s/'%(model.workdir,name)
+    os.chdir(dirname)
+    for file in os.listdir("."):
+        if fnmatch.fnmatch(file,"*.c"):
+            print "Compiling %s ..."%file,
+            prog_name=file[:-2]+".run"
+            process = subprocess.Popen(["rm", "-f", prog_name], shell=False, 
+                                        stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            exit_code = process.wait()
+            (std_out, std_err) = process.communicate()
+
+            process = subprocess.Popen(["gcc", file, "-lm", "-lpthread", 
+                                        "-lpvegas","-O2", "-o", prog_name], shell=False, 
+                                        stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            exit_code = process.wait()
+            (std_out, std_err) = process.communicate()
+
+            if exit_code <> 0 :
+                print "FAILED"
+            else: 
+                if len(std_err) == 0:
+                    print "OK"
+                else:
+                    print "CHECK"
+                    print std_err
+                        
+    return
+
 
 def save(name, graph, model, overwrite=True):
     dirname = '%s/%s/'%(model.workdir,name)
@@ -13,22 +104,28 @@ def save(name, graph, model, overwrite=True):
         if overwrite:
             file_list = os.listdir(dirname)
             for file in file_list:
-                os.remove(dirname+file)
+                if fnmatch.fnmatch(file,"*.c") or fnmatch.fnmatch(file,"*.run"):
+                    os.remove(dirname+file)
 
     jakob,subsvars = roperation.subs_vars(graph)
     cnt=0
     d,e=sympy.var('d e')
+    if "_nloops_orig" in graph.__dict__:
+        _nloops_orig=graph._nloops_orig
+    else:
+        _nloops_orig=graph.NLoops()
+    print "norm start"
+    norm=utils.series_f(utils.norm(graph.NLoops(),model.space_dim-e), e, model.target-_nloops_orig)
+    print norm
     for g in model.dTau(graph):
         roperation.strechMoments(g, model)
         print cnt, g
         det=roperation.det(g, model)
-        expr=(utils.norm(graph.NLoops(),d)*jakob*det*roperation.AvgByExtDir(roperation.expr(g,model))).subs(d, model.space_dim-e)
+        #expr=(jakob*det*roperation.AvgByExtDir(roperation.expr(g,model))).subs(d, model.space_dim-e)
+        expr=(norm*jakob*det*roperation.AvgByExtDir(roperation.expr(g,model))).subs(d, model.space_dim-e)
         strechs=roperation.find_strech_atoms(expr)
         eps_cnt=0
-        if "_nloops_orig" in graph.__dict__:
-            _nloops_orig=graph._nloops_orig
-        else:
-            _nloops_orig=graph.NLoops()
+
 
         for _expr in utils.series_lst(expr,e,model.target-_nloops_orig):
             integrand=roperation.export_subs_vars_pv(subsvars,strechs)
